backend/requirements.txt
langchain
langchain-community
langchain-experimental
langchain-openai
python-dotenv
fastapi
uvicorn
itsdangerous